pyproject.toml
[tool.poetry]
name = "photo-weather"
version = "0.1.0"
description = "Get most popular sunset pics from Flickr-->get time + location into--> get weather info--> infer best sunset weather --> alert me when it's optimal sunset weather"
authors = ["Will Cunningham"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
flickrapi = "^2.4.0"
python-dotenv = "^0.20.0"
noaa-sdk = "^0.1.21"

[tool.poetry.dev-dependencies]
ipykernel = "^6.15.1"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
